update clicker component

+++ client_gui/src/gui/window/game.rs
@@ -173,7 +173,7 @@ impl<'a> EventHandler for GameWindow<'a> {
         match e {
             Event::SFML(sfml::window::Event::MouseButtonReleased { button, x, y }) => {
                 if button == mouse::Button::Left {
-                    self.clicker.click(x as u32, y as u32);
+                    self.clicker.click(x, y);
                 }
             }
             Event::Network(NetworkEvent::PlayerJoined(e)) => {
