+++ client_gui/src/gui/window/settings.rs
@@ -120,7 +120,7 @@ impl<'a> EventHandler for SettingsWindow<'a> {
         match e {
             Event::SFML(sfml::window::Event::MouseButtonReleased { button, x, y }) => {
                 if button == mouse::Button::Left {
-                    self.clicker.click(x as u32, y as u32);
+                    self.clicker.click(x, y);
                 }
             }
             Event::UI(UIEvent::InputChanged { value }) => {
