+++ client_gui/src/gui/window/start.rs
@@ -85,7 +85,7 @@ impl<'a> EventHandler for StartWindow<'a> {
         match e {
             Event::SFML(sfml::window::Event::MouseButtonReleased { button, x, y }) => {
                 if button == mouse::Button::Left {
-                    self.clicker.click(x as u32, y as u32);
+                    self.clicker.click(x, y);
                 }
             }
             _ => {}
